+++ tests/snapshot_tests/test_snapshots.py
@@ -898,8 +898,10 @@ def test_unscoped_css(snap_compare) -> None:
 def test_big_buttons(snap_compare) -> None:
     assert snap_compare(SNAPSHOT_APPS_DIR / "big_button.py")
 
+
 def test_keyline(snap_compare) -> None:
     assert snap_compare(SNAPSHOT_APPS_DIR / "keyline.py")
+
 
 def test_button_outline(snap_compare):
     """Outline style rendered incorrectly when applied to a `Button` widget.
@@ -934,3 +936,14 @@ def test_vertical_max_height(snap_compare):
     """Test vertical max height takes border in to account."""
     assert snap_compare(SNAPSHOT_APPS_DIR / "vertical_max_height.py")
 
+
+def test_loading_indicator(snap_compare):
+    """Test loading indicator."""
+    assert snap_compare(SNAPSHOT_APPS_DIR / "loading.py", press=["space"])
+
+
+def test_loading_indicator_disables_widget(snap_compare):
+    """Test loading indicator."""
+    assert snap_compare(
+        SNAPSHOT_APPS_DIR / "loading.py", press=["space", "down", "down", "space"]
+    )